requirements.txt
Flask
pywebview
psutil
imageio
Nuitka[onefile]
